feat: floating-origin system for Earth-scale planets
Implement camera-relative rendering to eliminate f32 precision loss at
planetary distances. All simulation-relevant entities now use
WorldPosition(DVec3) as the source of truth; Transform becomes
render-space only (offset from RenderOrigin).

Phase 1A - Core types:
- WorldPosition(DVec3) component for true f64 positions
- RenderOrigin(DVec3) resource defining the f32 zero point
- RenderOriginShift resource for rebase events (512m threshold)
- FloatingOriginPlugin with rebase_origin and sync systems

Phase 1B - Camera migration:
- All camera systems (move, look, gravity, snap) use WorldPosition
- sync_camera_transform derives Transform from WorldPosition - RenderOrigin
- local_up_from_world_pos(DVec3) replaces local_up_from_position(Vec3)

Phase 1C - Chunk transform migration:
- v2_update_chunks reads camera WorldPosition for chunk selection
- v2_collect_results computes chunk positions relative to RenderOrigin
- v2_apply_origin_shift updates chunk Transforms on rebase
- CubeSphereCoord::world_transform_scaled_f64() returns DVec3

Phase 1D - Other systems migration:
- Save/load: serialize WorldPosition as f64 (PlayerSave position now [f64;3])
- Diagnostics: HUD and ECS dump display WorldPosition
- Map systems: local/global map and teleport use WorldPosition
- Audio: remains render-space (fine within 512m rebase threshold)

+++ src/persistence/types.rs
@@ -94,7 +94,7 @@ pub struct SaveGame {
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct PlayerSave {
-    pub position: [f32; 3],
+    pub position: [f64; 3],
     pub pitch: f32,
     pub yaw: f32,
     pub health_current: f32,
